Move fieldToString() from marcFieldSort.js to utils.js and make it exportable
